Restructure the integration tests
